feat(toast): integrate AppToastContainer and document helpers
- connected AppToastContainer in main.jsx for global toast rendering
- moved showSuccessToast and showErrorToast helpers into separate toastHelpers.js
- added JSDoc documentation with usage examples for toast helpers
