+++ src/main.jsx
@@ -4,6 +4,7 @@ import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { store } from './redux/store.js';
 import App from './App.jsx';
+import AppToastContainer from './components/common/AppToastContainer/AppToastContainer.jsx';
 import './styles/index.css';
 
 createRoot(document.getElementById('root')).render(
@@ -12,6 +13,7 @@ createRoot(document.getElementById('root')).render(
       <BrowserRouter>
         <App />
       </BrowserRouter>
+      <AppToastContainer />
     </Provider>
   </StrictMode>
 );
